Added main page with routing & search user ability, closes #25
Main page was added with a framework for future routing. Functionality for searching for users is added. This page layout is for demonstration, re-organization/modularization will be needed moving forwards. /Search path in express app was made to be able to return accept post requests and return a json response (without returning any extra information).

Co-Authored-By: loganmcmullen <47581369+loganmcmullen@users.noreply.github.com>