Changed database access structure #19
In the aim of reducing code-mess and removing the issue of the generic method to log into the database, I put login details into a separate "default.json" file. These two files should never really need to be altered and can be imported as an instance of our database cluster.