Search ability to find other users to follow #35
Search route should return a match for a searched user, which is the first step to being able to follow a user. After having this search result returned, the next step will be having a follow button next to the users name that follows the user. Included is a user-schema to assist with all routes.