added following users (closes #35) and follow list (closes #28)
Server: Two endpoints created for viewing all of the people who a user is following, and for adding a new follower.
Client: Added a page that displays all of the users who a person is following, and also the ability to add new users. This page uses the Navbar component and the new Following component.