Display 1 uploaded picture to the profile page  #61
A picture that has been uploaded is now able to be displayed to the profile page.

MUCH more features to come, just comitting a baseline for others to get started on their work.

This was done using JSX. basically implementing javascript functions into HTML tags so that once the html tags are called, it calls the function.

earlier commit included using GridFS and multer to upload the picture

lots of help from @Greg