Issue #61 Search and View other Profiles
Added functionality for searching for other users (in the Navbar) and viewing their profile. Setup this.props so that the history, location, and match URL parameters can be used.

Rather than further modify the current files, I chose to create new files for viewing other peoples profiles. While this creates more components and more pages, I felt it would be easier for others to follow along in this case.